requirements.txt
attrs==21.2.0
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
Flask==1.1.2
flask-expects-json==1.6.0
Flask-JSON==0.3.4
html2text==2020.1.16
idna==2.10
importlib-metadata==3.10.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
more-itertools==8.8.0
packaging==21.0
pip==21.2.4
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
pytest==6.2.4
python-dotenv==0.12.0
requests==2.25.1
setuptools==52.0.0.post20210125
six==1.16.0
toml==0.10.2
urllib3==1.26.6
Werkzeug==1.0.1
wheel==0.37.0
zipp==3.5.0
